requirements.txt
# ── Web Framework ────────────────────────────────────
flask>=3.0.0

# ── Hugging Face Inference API ───────────────────────
requests>=2.31.0
Pillow>=10.0.0
